requirements.txt
numpy>=1.15.1
onnx>=1.1.1
pycuda>=2017.1.1
opencv-python>=4.2.0.34
PyTurboJPEG>=1.4.1
scipy>=1.5.0
line_profiler
